backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
httpx==0.27.0
sqlalchemy==2.0.30
aiosqlite==0.20.0
alembic==1.13.1
redis==5.0.4
pydantic==2.7.1
pydantic-settings==2.2.1
python-dotenv==1.0.1
google-generativeai==0.5.4
solana==0.34.3
solders==0.21.0
base58==2.1.1
anyio==4.3.0
aiofiles==23.2.1
